Inline trait HasPageMetadata into AbstractPage
