+++ packages/framework/src/Contracts/AbstractPage.php
@@ -4,10 +4,14 @@ namespace Hyde\Framework\Contracts;
 
 use Hyde\Framework\Actions\SourceFileParser;
 use Hyde\Framework\Concerns\CanBeInNavigation;
-use Hyde\Framework\Concerns\HasPageMetadata;
+use Hyde\Framework\Helpers\Features;
+use Hyde\Framework\Helpers\Meta;
+use Hyde\Framework\Hyde;
 use Hyde\Framework\Models\FrontMatter;
+use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\Models\Route;
 use Hyde\Framework\Services\DiscoveryService;
+use Hyde\Framework\Services\RssFeedService;
 use Illuminate\Support\Collection;
 
 /**
@@ -23,7 +27,6 @@ use Illuminate\Support\Collection;
  */
 abstract class AbstractPage implements PageContract, CompilableContract
 {
-    use HasPageMetadata;
     use CanBeInNavigation;
 
     public static string $sourceDirectory;
@@ -158,4 +161,83 @@ abstract class AbstractPage implements PageContract, CompilableContract
 
     /** @inheritDoc */
     abstract public function compile(): string;
+
+    public function getCanonicalUrl(): string
+    {
+        return $this->getRoute()->getQualifiedUrl();
+    }
+
+    /**
+     * @return string[]
+     *
+     * @psalm-return list<string>
+     */
+    public function getDynamicMetadata(): array
+    {
+        $array = [];
+
+        if ($this->canUseCanonicalUrl()) {
+            $array[] = '<link rel="canonical" href="'.$this->getCanonicalUrl().'" />';
+        }
+
+        if (Features::sitemap()) {
+            $array[] = '<link rel="sitemap" type="application/xml" title="Sitemap" href="'.Hyde::url('sitemap.xml').'" />';
+        }
+
+        if (Features::rss()) {
+            $array[] = $this->makeRssFeedLink();
+        }
+
+        if (isset($this->title)) {
+            if ($this->hasTwitterTitleInConfig()) {
+                $array[] = '<meta name="twitter:title" content="'.$this->htmlTitle().'" />';
+            }
+            if ($this->hasOpenGraphTitleInConfig()) {
+                $array[] = '<meta property="og:title" content="'.$this->htmlTitle().'" />';
+            }
+        }
+
+        if ($this instanceof MarkdownPost) {
+            $array[] = "\n<!-- Blog Post Meta Tags -->";
+            foreach ($this->getMetadata() as $name => $content) {
+                $array[] = Meta::name($name, $content);
+            }
+            foreach ($this->getMetaProperties() as $property => $content) {
+                $array[] = Meta::property($property, $content);
+            }
+        }
+
+        return $array;
+    }
+
+    public function renderPageMetadata(): string
+    {
+        return Meta::render(
+            withMergedData: $this->getDynamicMetadata()
+        );
+    }
+
+    public function canUseCanonicalUrl(): bool
+    {
+        return Hyde::hasSiteUrl() && isset($this->identifier);
+    }
+
+    public function hasTwitterTitleInConfig(): bool
+    {
+        return str_contains(json_encode(config('hyde.meta', [])), 'twitter:title');
+    }
+
+    public function hasOpenGraphTitleInConfig(): bool
+    {
+        return str_contains(json_encode(config('hyde.meta', [])), 'og:title');
+    }
+
+    protected function makeRssFeedLink(): string
+    {
+        return sprintf(
+            '<link rel="alternate" type="application/rss+xml" title="%s" href="%s" />',
+            RssFeedService::getDescription(),
+            Hyde::url(RssFeedService::getDefaultOutputFilename())
+        );
+    }
 }
